requirements.txt
langchain
langchain-core
langchain-classic
langchain-community
langchain-ollama
langchain-google-genai
fastapi
uvicorn
pydantic
faiss-cpu
fastapi
uvicorn
supabase
dotenv
python-jose
langchain-groq
ragas
# -e .